requirements.txt
asgiref==3.5.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.4
django-braces==1.15.0
django-ckeditor==6.4.0
django-crispy-forms==1.14.0
django-js-asset==2.0.0
django-tinymce==3.4.0
idna==3.3
Pillow==9.1.0
psycopg2==2.9.3
pycryptodome==3.14.1
rave-python==1.2.15
requests==2.27.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
